+++ package.json
@@ -28,7 +28,6 @@
     "react-scripts": "5.0.1",
     "recoil": "^0.7.7",
     "socket.io-client": "^4.7.2",
-    "socket.io-client": "^4.7.2",
     "styled-components": "^6.1.0",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4",
