+++ src/components/openchat/invite/OpenchatInviteFriendItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Box, Tooltip, Typography } from '@mui/material';
 import { AddCircleOutline, RemoveCircleOutline } from '@mui/icons-material';
-import { OpenchatFriendWrap } from '../../styles/OpenchatStyle';
-import { User } from '../../types/User';
+import { OpenchatFriendWrap } from '../../../styles/OpenchatStyle';
+import { User } from '../../../types/User';
 
 interface OpenchatInviteFriendItemProps {
   user: User;
