+++ src/components/openchat/list/OpenchatCategory.tsx
@@ -1,30 +1,29 @@
 import React from 'react';
 import { Grid } from '@mui/material';
-import OpenchatMyItem from './OpenchatMyItem';
-import { Openchat } from '../../types/Openchat';
+import { ChatInfoWithId } from '../../../hooks/useQueryOpenchats';
+import OpenchatItem from './OpenchatItem';
 import OpenchatItemSkeleton from './OpenchatItemSkeleton';
 
-interface OpenchatMyProps {
+interface OpenchatCategoryProps {
   isQuering: boolean;
-  openchats?: Openchat[];
+  openchats?: ChatInfoWithId[];
 }
 
-function OpenchatMy({ isQuering, openchats }: OpenchatMyProps) {
+function OpenchatCategory({ isQuering, openchats }: OpenchatCategoryProps) {
   const count2 = ['skeleton1', 'skeleton2'];
   if (!openchats) return <div>채팅방이 없습니다.</div>;
   return (
     <Grid container spacing={2}>
       {!isQuering
         ? openchats.map((openchat) => (
-            <OpenchatMyItem key={openchat.id} openchat={openchat} />
+            <OpenchatItem key={openchat.id} openchat={openchat} />
           ))
         : count2.map((el) => <OpenchatItemSkeleton key={el} />)}
     </Grid>
   );
 }
 
-OpenchatMy.defaultProps = {
+OpenchatCategory.defaultProps = {
   openchats: null,
 };
-
-export default React.memo(OpenchatMy);
+export default React.memo(OpenchatCategory);
