+++ src/components/openchat/room/OpenchatNavUsers.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { Cycle, motion } from 'framer-motion';
+import { motion } from 'framer-motion';
 import OpenchatNavUserItem from './OpenchatNavUserItem';
-import { User } from '../../types/User';
+import { User } from '../../../types/User';
 import OpenchatNavInvite from './OpenchatNavInvite';
 
 // const variants = {
